Refactored based on feature
