+++ app/src/main/AndroidManifest.xml
@@ -4,7 +4,8 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"
+        tools:ignore="ScopedStorage" />
     <uses-permission android:name="android.permission.READ_MEDIA_IMAGES" />
 
     <application
@@ -12,7 +13,7 @@
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@drawable/logo_2"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
@@ -25,11 +26,8 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
-        
       <activity
             android:name=".feature_admin.ui.activity.AdminActivity"
-            android:name=".UpdateServiceActivity"
-
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -93,21 +91,21 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".presentation.register.RegisterActivity"
+            android:name=".feature_auth.ui.activity.RegisterActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".presentation.reset_password.ResetPasswordActivity"
+            android:name=".feature_auth.ui.activity.ResetPasswordActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".presentation.login.LoginActivity"
+            android:name=".feature_auth.ui.activity.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
