+++ app/src/main/java/com/wirajasa/wirajasabisnis/core/domain/repository/SellerRepository.kt
@@ -1,7 +1,7 @@
-package com.wirajasa.wirajasabisnis.data.repository
+package com.wirajasa.wirajasabisnis.core.domain.repository
 
-import com.wirajasa.wirajasabisnis.data.model.SellerApplication
-import com.wirajasa.wirajasabisnis.utility.NetworkResponse
+import com.wirajasa.wirajasabisnis.core.domain.model.SellerApplication
+import com.wirajasa.wirajasabisnis.core.utility.NetworkResponse
 import kotlinx.coroutines.flow.Flow
 
 interface SellerRepository {
