+++ app/src/main/java/com/wirajasa/wirajasabisnis/core/usecases/HandleException.kt
@@ -1,4 +1,4 @@
-package com.wirajasa.wirajasabisnis.usecases
+package com.wirajasa.wirajasabisnis.core.usecases
 
 import android.content.Context
 import com.google.firebase.auth.FirebaseAuthEmailException
@@ -7,10 +7,10 @@ import com.google.firebase.auth.FirebaseAuthInvalidUserException
 import com.wirajasa.wirajasabisnis.R
 import java.io.IOException
 
-class HandleException (private val context: Context) {
+class HandleException(private val context: Context, private val e: Exception) {
 
-    fun getMessage(exception: Exception): String {
-        return when (exception) {
+    operator fun invoke(): String {
+        return when (e) {
             is IOException -> context.getString(R.string.tv_connection_error)
             is FirebaseAuthEmailException -> context.getString(R.string.tv_invalid_email)
             is FirebaseAuthInvalidCredentialsException -> context.getString(R.string.tv_invalid_credential)
