+++ app/src/main/java/com/wirajasa/wirajasabisnis/di/AppModule.kt
@@ -8,7 +8,9 @@ import com.google.firebase.storage.ktx.storage
 import com.wirajasa.wirajasabisnis.feature_admin.domain.repository.AdminRepository
 import com.wirajasa.wirajasabisnis.feature_admin.domain.repository_impl.AdminRepositoryImpl
 import com.wirajasa.wirajasabisnis.core.crypto_pref.CryptoPref
-import com.wirajasa.wirajasabisnis.data.repository.*
+import com.wirajasa.wirajasabisnis.core.domain.repository.*
+import com.wirajasa.wirajasabisnis.feature_auth.domain.repository.AuthRepository
+import com.wirajasa.wirajasabisnis.feature_auth.domain.repository_impl.AuthRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
