+++ app/src/main/java/com/wirajasa/wirajasabisnis/feature_admin/domain/repository/AdminRepository.kt
@@ -1,7 +1,7 @@
 package com.wirajasa.wirajasabisnis.feature_admin.domain.repository
 
-import com.wirajasa.wirajasabisnis.data.model.SellerApplication
-import com.wirajasa.wirajasabisnis.utility.NetworkResponse
+import com.wirajasa.wirajasabisnis.core.domain.model.SellerApplication
+import com.wirajasa.wirajasabisnis.core.utility.NetworkResponse
 import kotlinx.coroutines.flow.Flow
 
 interface AdminRepository {
