+++ app/src/main/java/com/wirajasa/wirajasabisnis/feature_auth/ui/viewmodel/LoginViewModel.kt
@@ -1,13 +1,13 @@
-package com.wirajasa.wirajasabisnis.presentation.login
+package com.wirajasa.wirajasabisnis.feature_auth.ui.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import com.google.firebase.auth.FirebaseUser
-import com.wirajasa.wirajasabisnis.data.model.UserProfile
-import com.wirajasa.wirajasabisnis.data.repository.AuthRepository
-import com.wirajasa.wirajasabisnis.data.repository.UserRepository
-import com.wirajasa.wirajasabisnis.utility.NetworkResponse
+import com.wirajasa.wirajasabisnis.core.domain.model.UserProfile
+import com.wirajasa.wirajasabisnis.feature_auth.domain.repository.AuthRepository
+import com.wirajasa.wirajasabisnis.core.domain.repository.UserRepository
+import com.wirajasa.wirajasabisnis.core.utility.NetworkResponse
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Inject
