+++ app/src/main/java/com/wirajasa/wirajasabisnis/feature_auth/ui/viewmodel/ResetPasswordViewModel.kt
@@ -1,10 +1,10 @@
-package com.wirajasa.wirajasabisnis.presentation.reset_password
+package com.wirajasa.wirajasabisnis.feature_auth.ui.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
-import com.wirajasa.wirajasabisnis.data.repository.AuthRepository
-import com.wirajasa.wirajasabisnis.utility.NetworkResponse
+import com.wirajasa.wirajasabisnis.feature_auth.domain.repository.AuthRepository
+import com.wirajasa.wirajasabisnis.core.utility.NetworkResponse
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Inject
