+++ app/src/main/java/com/wirajasa/wirajasabisnis/presentation/profile/ProfileViewModel.kt
@@ -1,8 +1,8 @@
 package com.wirajasa.wirajasabisnis.presentation.profile
 
 import androidx.lifecycle.ViewModel
-import com.wirajasa.wirajasabisnis.data.model.UserProfile
-import com.wirajasa.wirajasabisnis.data.repository.UserRepository
+import com.wirajasa.wirajasabisnis.core.domain.model.UserProfile
+import com.wirajasa.wirajasabisnis.core.domain.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
