+++ app/src/main/java/com/wirajasa/wirajasabisnis/presentation/service/AddingServiceViewModel.kt
@@ -4,13 +4,9 @@ import android.net.Uri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
-import com.google.firebase.auth.FirebaseUser
-import com.wirajasa.wirajasabisnis.data.model.SellerApplication
-import com.wirajasa.wirajasabisnis.data.model.UserProfile
-import com.wirajasa.wirajasabisnis.data.repository.AuthRepository
-import com.wirajasa.wirajasabisnis.data.repository.ProductRepository
-import com.wirajasa.wirajasabisnis.data.repository.UserRepository
-import com.wirajasa.wirajasabisnis.utility.NetworkResponse
+import com.wirajasa.wirajasabisnis.core.domain.model.SellerApplication
+import com.wirajasa.wirajasabisnis.core.domain.repository.ProductRepository
+import com.wirajasa.wirajasabisnis.core.utility.NetworkResponse
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flowOn
