+++ app/src/main/java/com/wirajasa/wirajasabisnis/presentation/service/UpdateServiceActivity.kt
@@ -5,8 +5,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import com.bumptech.glide.Glide
-import com.wirajasa.wirajasabisnis.R
-import com.wirajasa.wirajasabisnis.data.model.ServicePost
+import com.wirajasa.wirajasabisnis.core.domain.model.ServicePost
 import com.wirajasa.wirajasabisnis.databinding.ActivityUpdateServiceBinding
 
 class UpdateServiceActivity : AppCompatActivity() {
@@ -18,7 +17,7 @@ class UpdateServiceActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         post = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            intent.getParcelableExtra(DetailServiceActivity.EXTRA_SERVICE_POST,ServicePost::class.java)
+            intent.getParcelableExtra(DetailServiceActivity.EXTRA_SERVICE_POST, ServicePost::class.java)
         } else {
             intent.getParcelableExtra<ServicePost>(DetailServiceActivity.EXTRA_SERVICE_POST) as ServicePost
         }
