+++ app/src/main/java/com/wirajasa/wirajasabisnis/ui/dashboard/BuyerDashboardActivity.kt
@@ -3,11 +3,9 @@ package com.wirajasa.wirajasabisnis.ui.dashboard
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.get
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.wirajasa.wirajasabisnis.R
 import com.wirajasa.wirajasabisnis.adapter.ProductAdapter
-import com.wirajasa.wirajasabisnis.data.model.Product
+import com.wirajasa.wirajasabisnis.core.domain.model.Product
 import com.wirajasa.wirajasabisnis.databinding.ActivityBuyerDashboardBinding
 
 class BuyerDashboardActivity : AppCompatActivity() {
