+++ app/src/main/java/com/wirajasa/wirajasabisnis/ui/dashboard/SellerDashboardViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.wirajasa.wirajasabisnis.data.model.Product
+import com.wirajasa.wirajasabisnis.core.domain.model.Product
 import kotlinx.coroutines.launch
 
 enum class SellerProductApiStatus {LOADING, SUCCESS, ERROR}
