+++ app/src/main/java/com/wirajasa/wirajasabisnis/ui/seller/ui/dashboard/DashboardViewModel.kt
@@ -1,14 +1,13 @@
 package com.wirajasa.wirajasabisnis.ui.seller.ui.dashboard
 
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
-import com.wirajasa.wirajasabisnis.data.model.ServicePost
-import com.wirajasa.wirajasabisnis.data.model.UserProfile
-import com.wirajasa.wirajasabisnis.data.repository.ProductRepository
-import com.wirajasa.wirajasabisnis.data.repository.UserRepository
-import com.wirajasa.wirajasabisnis.utility.NetworkResponse
+import com.wirajasa.wirajasabisnis.core.domain.model.ServicePost
+import com.wirajasa.wirajasabisnis.core.domain.model.UserProfile
+import com.wirajasa.wirajasabisnis.core.domain.repository.ProductRepository
+import com.wirajasa.wirajasabisnis.core.domain.repository.UserRepository
+import com.wirajasa.wirajasabisnis.core.utility.NetworkResponse
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flowOn
