+++ app/src/main/java/com/wirajasa/wirajasabisnis/ui/seller/ui/home/ProfileViewModel.kt
@@ -1,10 +1,10 @@
 package com.wirajasa.wirajasabisnis.ui.seller.ui.home
 
 import androidx.lifecycle.ViewModel
-import com.wirajasa.wirajasabisnis.data.model.SellerApplication
-import com.wirajasa.wirajasabisnis.data.model.UserProfile
-import com.wirajasa.wirajasabisnis.data.repository.SellerRepository
-import com.wirajasa.wirajasabisnis.data.repository.UserRepository
+import com.wirajasa.wirajasabisnis.core.domain.model.SellerApplication
+import com.wirajasa.wirajasabisnis.core.domain.model.UserProfile
+import com.wirajasa.wirajasabisnis.core.domain.repository.SellerRepository
+import com.wirajasa.wirajasabisnis.core.domain.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
