--- app/src/main/java/com/wirajasa/wirajasabisnis/utility/Utils.kt
@@ -1,36 +1,0 @@
-package com.wirajasa.wirajasabisnis.utility
-
-import android.content.ContentResolver
-import android.content.Context
-import android.net.Uri
-import android.os.Environment
-import java.io.*
-import java.text.SimpleDateFormat
-import java.util.*
-
-private const val FILENAME_FORMAT = "dd-MMM-yyyy"
-private const val IMAGE_TYPE = ".jpg"
-val timeStamp: String = SimpleDateFormat(
-    FILENAME_FORMAT,
-    Locale.US
-).format(System.currentTimeMillis())
-
-fun createTempFile(context: Context): File {
-    val storageDir: File? = context.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
-    return File.createTempFile(timeStamp, IMAGE_TYPE, storageDir)
-}
-
-fun uriToFile(selectedImg: Uri, context: Context): File {
-    val contentResolver: ContentResolver = context.contentResolver
-    val myFile = createTempFile(context)
-
-    val inputStream = contentResolver.openInputStream(selectedImg) as InputStream
-    val outputStream: OutputStream = FileOutputStream(myFile)
-    val buf = ByteArray(1024)
-    var len: Int
-    while (inputStream.read(buf).also { len = it } > 0) outputStream.write(buf, 0, len)
-    outputStream.close()
-    inputStream.close()
-
-    return myFile
-}
