+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/orange"
-    tools:context=".presentation.register.RegisterActivity">
+    tools:context=".feature_auth.ui.activity.RegisterActivity">
 
     <com.google.android.material.card.MaterialCardView
         android:layout_width="match_parent"
