+++ app/src/main/res/layout/activity_reset_password.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/orange"
-    tools:context=".presentation.reset_password.ResetPasswordActivity">
+    tools:context=".feature_auth.ui.activity.ResetPasswordActivity">
 
     <com.google.android.material.card.MaterialCardView
         android:layout_width="match_parent"
